# optimize photos search sql request