@lastzero 2021-01-10
Excellent catch! 

Did you also test this with MariaDB? Haven't observed such slow queries over here, and I have much more files and photos.