@ya-aleksko 2021-01-10
> Did you also test this with MariaDB? Haven't observed such slow queries over here, and I have much more files and photos.

Oh, sorry, I checked requests using mariadb from photoprism docker container. My photoprism and mariadb is installed on old and quite slow computer (Intel Atom D525 1.8GHz 3Gb RAM), may be it`s a reason of so bad timings. May I ask you check my queries (old and new) on your installation and compare timings?